requirements.txt
requests==2.32.5
python-dotenv==1.1.1